requirements.txt
telethon==1.34.0
pytgcalls==0.9.7
python-dotenv==1.0.0
ffmpeg-python==0.2.0
tgcrypto